Comp_huff

Description

This module defines the necessary functionality for huffman compression Includes functions for compression alongwith some encode functions

Compression Related function

newHuffTree :: Ord a => [(a, Integer)] -> HuffTree a

Function accepts a list consisting of a set of (symbol ,ocurrence) pair,builds a Huffman tree representation

newHuffTree' :: Ord a => BinomialHeap (HuffTree a) -> HuffTree a

Function accepts a BinomialHeap which is a set of Huffman trees and returns a single Huffman tree consisting of the first values

encodeHuff :: HuffTree Byte -> [Bool]

Function which reads input.If it comes across a True, then it can be first True (either at the beginning of the code or after a False) then the next bit is part of the tree (represented inversely).Otherwise, it is considered as finalized the code. If a False is read, then the next 8 bits, correspond to a symbol.

huffComp :: Ord a => [a] -> (HuffTree a, [Bool])

The function accepts a list of symbols, return a tuple(elem ,binarylist).elem is the first of given symbols. binarylist is a list where symbols have been replaced by binary and concatenated

huffCompress :: [Byte] -> [Byte]

The function changes a byte list into another byte list , compressing it The last True is added in order to have an end of file character (All of the False found before the last True are ignored by the decoder)

getInstances :: Eq a => [a] -> [(a, Integer)] -> [(a, Integer)]

Function accepts a symbol list and returns a list of tuples (symbol ,occurrence)