Huffman_shared

Description

This module defines the necessary functionality common to Huffman compression and decompression modules.

Type Byte

type Byte = Word8

A Byte is an 8 bit word

Type HuffTree

data HuffTree a

Huffman Tree type definition

Common Operations on Huffman Compression and Decompression

readInputFile :: String -> IO [Byte]

Function that accepts a (relative) file path and returns the content of the file as a list of bytes (Byte)

printHufFile :: String -> [Byte] -> IO ()

Function that accepts a (relative) file path and a list of bytes (Byte) and prints the bytes in the input file.

byteToBits :: Byte -> [Bool]

Function that accepts a byte (@Byte) and outputs the corresponding list of bits.

bitsToBytes :: [Bool] -> [Byte]

Function that accepts a list of booleans (bits) and outputs a list of bytes

frequency :: HuffTree huf -> Integer

Function that accepts a Huffman tree and gets the accumulated frequency in the root of the tree.

mergeHuffTrees :: HuffTree huf -> HuffTree huf -> HuffTree huf

Function that accepts two Huffman trees and returns a new tree with the original ones as children of this tree. The accumulated frequency of this tree is the sum of the frequencies of the two children.

constructHeap :: Ord a => [(a, Integer)] -> BinomialHeap (HuffTree a)

Function that accepts a list of elements with its corresponding frequencies and constructs a Binomial Heap of associated Huffman trees.

makeHuffTree :: [(a, [Bool])] -> HuffTree a

Function whcih accepts a list of tuples with a set of symbols, next totheir binary coding (False represents zero and True the one). It Builds a Huffman Tree such that these association they could have been generated.The accumulated frequency will be assigned to zero (0) for all the leaves and branches of it

retSym :: HuffTree a -> [Bool] -> [a]

Accepts a Huffman tree and a list that represents a chain of bits, returns a list of symbols that corresponds to the decoding of the tree's information It is called by huffDecompress

getTuples :: Ord a => HuffTree a -> [(a, [Bool])]

Function accepts Huffman Tree and outputs a list of tuples with entry as (Symbol ,Coding Binary)