Lzw_Compress

Description

This module defines the necessary functionality for LZW compression Includes functions for compression alongwith some encode functions

Class Lzw_Compress

compChar :: [Word8] -> CompDict -> ([Word8], Abbreviation)

Function that accepts a string and a (compression) dictionary and ouptus the abbreviation (based on the string ) & remaining part ofthe string.

compressString :: [Word8] -> [Abbreviation]

Function which accepts a list of Word8s which are compressed into a list of abbrevations and outputs this new list

compressByteString :: ByteString -> ByteString

Function which accepts a byteString (uncompressed file) and outputs a byteString (compressed file).