Lzw_shared

Description

This module defines the necessary functionality common to Lzw compression and decompression modules.

Class Lzw_shared

type DecompDict = Map Id Abbreviation

Type of Dictionary used for compression

type CompDict = Map Abbreviation Id

Type of Dictionary used for compression

type Abbreviation = (Id, Word8)

Abbreviation is a tuple of Id and a string to be mapped

type Id = Word16

Index in the table

initCompDict :: CompDict

Function to initialize the compression dictionary

initDecompDict :: DecompDict

Function to initialize the compression dictionary

printLzwFile :: FilePath -> ByteString -> IO ()

Function to write content to output file