BinomialHeap

Description

Data Structure: Binomial Heap

For Huffman Compression

Class BinomialHeap

data BinomialTree a

Constructors

Node Int a [BinomialTree a]

data BinomialHeap a

Constructors

BH [BinomialTree a]

height :: BinomialTree a -> Int

Get the height of the binomial tree

root :: BinomialTree a -> a

Get the root of the binomial tree

linkTrees :: Ord a => BinomialTree a -> BinomialTree a -> BinomialTree a

Function that accepts two trees of height h | Returns a tree of height (h+1) | containing same elements.

insertTree :: Ord a => BinomialTree a -> [BinomialTree a] -> [BinomialTree a]

Function that accepts a list of trees and a tree | Inserts this tree at corresponding position in the list | Returns the modified list

insert :: Ord a => a -> BinomialHeap a -> BinomialHeap a

Function to insert a Node instance in the BinomialHeap

mergeTrees :: Ord a => [BinomialTree a] -> [BinomialTree a] -> [BinomialTree a]

Merges two BinomialTree instances and returns one

deleteMinimumTree :: Ord a => [BinomialTree a] -> (BinomialTree a, [BinomialTree a])

Auxillary function that accepts a BinomialTree and removes the minimum. | Returns minimum and the remainder tree.

findMinimum :: Ord a => BinomialHeap a -> a

Function that accepts a BinomialHeap and returns the minimum element.

deleteMinimum :: Ord a => BinomialHeap a -> BinomialHeap a

Function that accepts a BinomialHeap and eliminates the minimum