Decomp_huff

Description

This module defines the necessary functionality for HUFF decompression Includes functions for decompression alongwith some some decode functions

Decompression Related function

decodeHuff :: [Bool] -> (HuffTree Byte, [Bool])

Function reads the encoding of a Huffman Tree and reconstructs the Tree from which it was generated, separating the rest of the binary data.

huffDecomp :: HuffTree a -> [Bool] -> [a]

Given a Huffman tree and a list that represents a chain of bits, returns a list of symbols that corresponds to the decoding of said chain with the tree's information

huffDecompress :: [Byte] -> [Byte]

Function transforms a byte list into another one, decompressing it