Lzw_Decompress

Description

This module defines the necessary functionality for Lzw decompression Includes functions for decompression alongwith some some decode functions

Class Lzw_Decompress

getAbbr :: Get Abbreviation

Function returns Get of an Abbreviation

getAbbrList :: Get [Abbreviation]

Function to get a list of abbreviations from a ByteString (uncompressed file)

decodeAbbreviation :: Abbreviation -> DecompDict -> [Word8]

Function which accepts an abbreviation and a (decompression) dictionary and outputs the string based on the dictionary

decompressString :: [Abbreviation] -> [Word8]

Function that accepts a list of entries and generates corresponding list of words